PR feedback: remove unneeded interface layers and add org ID and realm to request
